+++ client/src/components/SignIn/Label.module.scss
@@ -1,0 +1,21 @@
+@use '../../styles/color' as C;
+@use '../../styles/utils' as U;
+@use '../../styles/typography' as T;
+
+.label {
+  @include T.black-regular-14;
+  @include U.flex-c(column, flex-start, flex-start);
+  gap: 12px;
+
+  input {
+    padding-left: 16px;
+    border: 1px solid C.$gray300;
+    border-radius: 6px;
+    @include U.size(300px, 40px);
+  }
+}
+
+.passwordInput {
+  @include U.flex-c(row, flex-start, center);
+  position: relative;
+}
