+++ client/src/pages/SignIn.module.scss
@@ -8,24 +8,6 @@
   gap: 24px;
 }
 
-.label {
-  @include T.black-regular-14;
-  @include U.flex-c(column, flex-start, flex-start);
-  gap: 12px;
-
-  input {
-    padding-left: 16px;
-    border: 1px solid C.$gray300;
-    border-radius: 6px;
-    @include U.size(300px, 40px);
-  }
-}
-
-.passwordInput {
-  @include U.flex-c(row, flex-start, center);
-  position: relative;
-}
-
 .btnBox {
   @include U.flex-c(row, flex-start, center);
   @include T.blue-regular-16;
